requirements.txt
pinecone-client==3.0.0
openai==1.12.0
python-dotenv==1.0.0
sentence-transformers==2.2.2
torch>=1.9.0
